+++ packages/wrapper/test/editorAppBase.test.ts
@@ -4,7 +4,6 @@
  * ------------------------------------------------------------------------------------------ */
 
 import { describe, expect, test } from 'vitest';
-import { isCodeUpdateRequired, isModelUpdateRequired, ModelUpdateType } from 'monaco-editor-wrapper';
 import { createWrapperConfigClassicApp, createWrapperConfigExtendedApp } from './helper.js';
 
 describe('Test EditorAppBase', () => {
@@ -17,45 +16,6 @@ describe('Test EditorAppBase', () => {
     test('extended type: empty EditorAppConfigExtended', () => {
         const wrapperConfig = createWrapperConfigExtendedApp();
         expect(wrapperConfig.editorAppConfig.$type).toBe('extended');
-    });
-
-    test('isCodeUpdateRequired', () => {
-        let codeUpdateType = isCodeUpdateRequired({ main: { text: '', fileExt: 'js' } }, { main: { text: '', fileExt: 'js' } });
-        expect(codeUpdateType).toBe(ModelUpdateType.NONE);
-
-        codeUpdateType = isCodeUpdateRequired({}, {});
-        expect(codeUpdateType).toBe(ModelUpdateType.NONE);
-
-        codeUpdateType = isCodeUpdateRequired({ main: { text: 'bar', fileExt: 'js' } }, { main: { text: '', fileExt: 'js' } });
-        expect(codeUpdateType).toBe(ModelUpdateType.CODE);
-
-        codeUpdateType = isCodeUpdateRequired({ main: { text: '', fileExt: 'js' } }, { main: { text: 'bar', fileExt: 'js' } });
-        expect(codeUpdateType).toBe(ModelUpdateType.CODE);
-
-        codeUpdateType = isCodeUpdateRequired({}, { main: { text: 'bar', fileExt: 'js' } });
-        expect(codeUpdateType).toBe(ModelUpdateType.CODE);
-
-        codeUpdateType = isCodeUpdateRequired({ main: { text: 'bar', fileExt: 'js' } }, {});
-        expect(codeUpdateType).toBe(ModelUpdateType.CODE);
-    });
-
-    test('isModelUpdateRequired', () => {
-        const config = createWrapperConfigClassicApp().editorAppConfig;
-
-        let modelUpdateType = isModelUpdateRequired(config.codeResources, { main: { text: '', fileExt: 'js' } });
-        expect(modelUpdateType).toBe(ModelUpdateType.NONE);
-
-        modelUpdateType = isModelUpdateRequired(config.codeResources, { main: { text: 'bar', fileExt: 'js' } });
-        expect(modelUpdateType).toBe(ModelUpdateType.CODE);
-
-        modelUpdateType = isModelUpdateRequired(config.codeResources, { main: { text: 'bar', uri: 'file:///bar.js' } });
-        expect(modelUpdateType).toBe(ModelUpdateType.CODE_AND_MODEL);
-
-        modelUpdateType = isModelUpdateRequired(config.codeResources, { original: { text: 'bar', fileExt: 'python' } });
-        expect(modelUpdateType).toBe(ModelUpdateType.CODE_AND_MODEL);
-
-        modelUpdateType = isModelUpdateRequired(config.codeResources, {});
-        expect(modelUpdateType).toBe(ModelUpdateType.CODE_AND_MODEL);
     });
 
 });