+++ packages/client/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 All notable changes to this npm module are documented in this file.
 
-## [9.0.0-next.2] - 2024-10-10
+## [9.0.0-next.3] - 2024-10-11
 
 - Support all arguments for monaco-vscode-api `initialize` [#756](https://github.com/TypeFox/monaco-languageclient/pull/756)
 - Update to monaco-vscode-api 9.0.x [#749](https://github.com/TypeFox/monaco-languageclient/pull/749)
-  - Updated all `@codingame/monaco-vscode` packages to `10.0.1`.
+  - Updated all `@codingame/monaco-vscode` packages to `10.0.2`.
   - Enhancements to logging
 - monaco-languageclient config improvement, wrapper+languageclientwrapper improvements [#741](https://github.com/TypeFox/monaco-languageclient/pull/741)
   - Pass MessageTransports directly